Move JupyterCellRenderer
